new sync only and separated fflate modules for smaller builds without advanced treeshaking
