+++ examples/jsm/loaders/3MFLoader.js
@@ -19,7 +19,7 @@ import {
 	TextureLoader,
 	SRGBColorSpace
 } from 'three';
-import { unzipSync } from '../libs/fflate.module.js';
+import { unzipSync } from '../libs/fflate.unzip.module.js';
 
 const COLOR_SPACE_3MF = SRGBColorSpace;
