+++ examples/webgl_rendertarget_texture2darray.html
@@ -113,7 +113,7 @@
 			import * as THREE from 'three';
 
 			import Stats from 'three/addons/libs/stats.module.js';
-			import { unzipSync } from 'three/addons/libs/fflate.module.js';
+			import { unzipSync } from 'three/addons/libs/fflate.unzip.module.js';
 			import { GUI } from 'three/addons/libs/lil-gui.module.min.js';
 
 			const DIMENSIONS = {
