+++ examples/webgl_texture2darray.html
@@ -69,7 +69,7 @@
 			import * as THREE from 'three';
 
 			import Stats from 'three/addons/libs/stats.module.js';
-			import { unzipSync } from 'three/addons/libs/fflate.module.js';
+			import { unzipSync } from 'three/addons/libs/fflate.unzip.module.js';
 
 			let camera, scene, mesh, renderer, stats;
