+++ examples/webgpu_rendertarget_2d-array_3d.html
@@ -61,7 +61,7 @@
 
 			import { TextureHelper } from 'three/addons/helpers/TextureHelperGPU.js';
 			import { OrbitControls } from 'three/addons/controls/OrbitControls.js';
-			import { unzipSync } from 'three/addons/libs/fflate.module.js';
+			import { unzipSync } from 'three/addons/libs/fflate.unzip.module.js';
 
 			let renderer;
 			let views = [];
