+++ examples/webgpu_textures_2d-array.html
@@ -41,7 +41,7 @@
 			import * as THREE from 'three/webgpu';
 			import { texture, uv, time, oscTriangle } from 'three/tsl';
 
-			import { unzipSync } from 'three/addons/libs/fflate.module.js';
+			import { unzipSync } from 'three/addons/libs/fflate.unzip.module.js';
 
 			//
