+++ test/unit/addons/exporters/USDZExporter.tests.js
@@ -1,7 +1,7 @@
 import { BoxGeometry, Mesh, MeshStandardMaterial, Scene, SphereGeometry } from 'three';
 import { USDZExporter } from '../../../../examples/jsm/exporters/USDZExporter.js';
 import { USDLoader } from '../../../../examples/jsm/loaders/USDLoader.js';
-import { unzipSync, strFromU8 } from '../../../../examples/jsm/libs/fflate.module.js';
+import { unzipSync, strFromU8 } from '../../../../examples/jsm/libs/fflate.unzip.module.js';
 
 function isValidUSDA( usda ) {
