+++ utils/vendor-fflate.js
@@ -1,0 +1,298 @@
+import { existsSync, readFileSync, writeFileSync } from 'node:fs';
+import { join, relative, resolve } from 'node:path';
+
+const VERSION = '0.8.2';
+const SOURCE_URL = `https://unpkg.com/fflate@${ VERSION }/esm/browser.js`;
+const OUT_DIR = 'examples/jsm/libs';
+
+const args = process.argv.slice( 2 );
+
+function argValue( name ) {
+
+	const index = args.indexOf( name );
+	return index === - 1 ? null : args[ index + 1 ];
+
+}
+
+function hasArg( name ) {
+
+	return args.includes( name );
+
+}
+
+async function readSource() {
+
+	const input = argValue( '--input' );
+	const fflateSrc = argValue( '--fflate-src' );
+
+	if ( input ) {
+
+		return {
+			path: resolve( input ),
+			code: readFileSync( input, 'utf8' )
+		};
+
+	}
+
+	if ( fflateSrc ) {
+
+		const sourcePath = join( fflateSrc, 'esm/browser.js' );
+
+		if ( ! existsSync( sourcePath ) ) {
+
+			throw new Error( `Expected built fflate browser ESM at ${ sourcePath }. Run fflate's build first or pass --input.` );
+
+		}
+
+		return {
+			path: resolve( sourcePath ),
+			code: readFileSync( sourcePath, 'utf8' )
+		};
+
+	}
+
+	const response = await fetch( SOURCE_URL );
+
+	if ( ! response.ok ) {
+
+		throw new Error( `Could not fetch ${ SOURCE_URL }: ${ response.status } ${ response.statusText }` );
+
+	}
+
+	return {
+		path: SOURCE_URL,
+		code: await response.text()
+	};
+
+}
+
+function section( code, start, end ) {
+
+	const startIndex = code.indexOf( start );
+
+	if ( startIndex === - 1 ) {
+
+		throw new Error( `Could not find start marker: ${ start }` );
+
+	}
+
+	const endIndex = code.indexOf( end, startIndex );
+
+	if ( endIndex === - 1 ) {
+
+		throw new Error( `Could not find end marker: ${ end }` );
+
+	}
+
+	return code.slice( startIndex, endIndex );
+
+}
+
+function stripExports( code ) {
+
+	return code.replace( /^export /gm, '' );
+
+}
+
+function from( code, start ) {
+
+	const startIndex = code.indexOf( start );
+
+	if ( startIndex === - 1 ) {
+
+		throw new Error( `Could not find start marker: ${ start }` );
+
+	}
+
+	return code.slice( startIndex );
+
+}
+
+function banner( sourcePath, name ) {
+
+	const displaySourcePath = sourcePath.startsWith( 'http' ) ? sourcePath : relative( process.cwd(), sourcePath ) || sourcePath;
+
+	return `/*!
+fflate - fast JavaScript compression/decompression
+<https://101arrowz.github.io/fflate>
+Licensed under MIT. https://github.com/101arrowz/fflate/blob/master/LICENSE
+version ${ VERSION }
+
+This ${ name } entrypoint is generated by utils/vendor-fflate.js from:
+${ displaySourcePath }
+*/
+
+`;
+
+}
+
+function writeGenerated( file, content ) {
+
+	writeFileSync( join( OUT_DIR, file ), content );
+	console.log( `Wrote ${ join( OUT_DIR, file ) }` );
+
+}
+
+const { path: sourcePath, code } = await readSource();
+
+const aliasesToError = stripExports( section( code, '// aliases for shorter compressed code', '// expands raw DEFLATE data' ) );
+const byteHelpers = section( code, '// read 2 bytes', '// gzip header' );
+const textPrelude = stripExports( section( code, '// text encoder', '/**\n * Streaming UTF-8 decoding' ) );
+const strToU8 = stripExports( section( code, '/**\n * Converts a string into a Uint8Array', '/**\n * Converts a Uint8Array to a string' ) );
+const strFromU8 = stripExports( section( code, '/**\n * Converts a Uint8Array to a string', ';\n// deflate bit flag' ) );
+
+const shared = `${ banner( sourcePath, 'shared' ) }${ aliasesToError }
+// empty
+var et = /*#__PURE__*/ new u8(0);
+${ byteHelpers }
+${ textPrelude }
+${ strToU8 }
+${ strFromU8 };
+
+export {
+	u8,
+	u16,
+	i32,
+	fleb,
+	fdeb,
+	clim,
+	fl,
+	revfl,
+	fd,
+	revfd,
+	rev,
+	flt,
+	fdt,
+	flm,
+	flrm,
+	fdm,
+	fdrm,
+	max,
+	bits,
+	bits16,
+	shft,
+	slc,
+	FlateErrorCode,
+	err,
+	hMap,
+	b2,
+	b4,
+	b8,
+	wbytes,
+	et,
+	strToU8,
+	strFromU8
+};
+`;
+
+const inflateCore = section( code, '// expands raw DEFLATE data', '// starting at p, write the minimum number of bits' );
+const unzipWrappers = section( code, '// gzip start', 'function StrmOpt' );
+const inflateSync = section( code, '/**\n * Expands DEFLATE data', '// before you yell at me' );
+const gunzipSync = section( code, '/**\n * Expands GZIP data', '/**\n * Streaming Zlib compression' );
+const unzlibSync = section( code, '/**\n * Expands Zlib data', '// Default algorithm for compression' );
+const zipReadHelpers = section( code, '// skip local zip header', '// extra field length' );
+const unzipSync = from( code, '/**\n * Synchronously decompresses a ZIP archive' );
+
+const unzip = `${ banner( sourcePath, 'decompression' ) }import {
+	u8,
+	fleb,
+	fdeb,
+	clim,
+	fl,
+	fd,
+	flrm,
+	fdrm,
+	max,
+	bits,
+	bits16,
+	shft,
+	slc,
+	err,
+	hMap,
+	b2,
+	b4,
+	b8,
+	strFromU8
+} from './fflate.shared.module.js';
+
+export { FlateErrorCode, strFromU8 } from './fflate.shared.module.js';
+
+${ inflateCore }
+${ unzipWrappers }
+${ inflateSync }
+${ gunzipSync }
+${ unzlibSync }
+${ zipReadHelpers }
+${ unzipSync }
+`;
+
+const compressionCore = section( code, '// starting at p, write the minimum number of bits', '// worker clone' );
+const zipWriteWrappers = section( code, '// gzip header', '// zlib start' );
+const deflateSync = section( code, '/**\n * Compresses data with DEFLATE without any wrapper', '/**\n * Streaming DEFLATE decompression' );
+const gzipSync = section( code, '/**\n * Compresses data with GZIP', '/**\n * Streaming single or multi-member GZIP decompression' );
+const zlibSync = section( code, '/**\n * Compress data with Zlib', '/**\n * Streaming Zlib decompression' );
+const fltn = section( code, '// flatten a directory structure', '// text encoder' );
+const zipWriteHelpers = section( code, '// extra field length', '/**\n * A pass-through stream to keep data uncompressed in a ZIP archive.' );
+const zipSync = section( code, '/**\n * Synchronously creates a ZIP file.', '/**\n * Streaming pass-through decompression for ZIP archives' );
+
+const zip = `${ banner( sourcePath, 'compression/archive' ) }import {
+	u8,
+	u16,
+	i32,
+	fleb,
+	fdeb,
+	clim,
+	revfl,
+	revfd,
+	rev,
+	flt,
+	fdt,
+	flm,
+	fdm,
+	hMap,
+	shft,
+	slc,
+	err,
+	wbytes,
+	strToU8
+} from './fflate.shared.module.js';
+
+export { FlateErrorCode, strToU8 } from './fflate.shared.module.js';
+
+${ compressionCore }
+${ zipWriteWrappers }
+${ deflateSync }
+${ gzipSync }
+${ zlibSync }
+${ fltn }
+${ zipWriteHelpers }
+${ zipSync }
+`;
+
+const barrel = `${ banner( sourcePath, 'compatibility' ) }export {
+	FlateErrorCode,
+	gunzipSync,
+	inflateSync,
+	strFromU8,
+	unzipSync,
+	unzlibSync
+} from './fflate.unzip.module.js';
+
+export {
+	deflateSync,
+	gzipSync,
+	strToU8,
+	zlibSync,
+	zipSync
+} from './fflate.zip.module.js';
+`;
+
+if ( ! hasArg( '--check' ) ) {
+
+	writeGenerated( 'fflate.module.js', barrel );
+	writeGenerated( 'fflate.shared.module.js', shared );
+	writeGenerated( 'fflate.unzip.module.js', unzip );
+	writeGenerated( 'fflate.zip.module.js', zip );
+
+}
+
